requirements.txt
streamlit==1.31.1
snowflake-connector-python==3.7.1
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
cryptography==42.0.2
Pillow==10.2.0
scipy==1.12.0
